requirements.txt
# Core ML/AI packages
torch>=2.0.0
transformers>=4.30.0
datasets>=2.13.0
peft>=0.4.0
accelerate>=0.20.0
sentencepiece>=0.1.99

# API and Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Frontend Framework
streamlit>=1.28.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0

# Logging and Monitoring
python-json-logger>=2.0.7

# Development
black>=23.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: for cloud deployment
# gunicorn>=21.2.0  # For production WSGI server
